Move queryoptimiser implementation into base class
